feat: rename env package to environment
