+++ server/api/lib/environment/environment.go
@@ -1,4 +1,4 @@
-package env
+package environment
 
 import (
 	"fmt"
